requirements.txt
numpy==1.18.1
pandas==1.0.1
requests==2.23.0
tqdm==4.43.0
regex==2020.2.20
torch==1.4.0
transformers==2.5.1
scipy==1.4.1
scikit-learn==0.22.2.post1
tensorboardX==2.0
wandb==0.8.29
seaborn==0.10.0
GoogleDriveDownloader
sentence-transformers==0.2.5.1
progressbar2==3.50.1